+++ packages/common/store/reducers/index.js
@@ -1,4 +1,4 @@
-export { default as wallet } from '@wallet/store/reducer';
+export { default as wallet } from '@wallet/store/reducers/reducer';
 export { default as blocks } from '@block/store/reducer';
 export { default as bookmarks } from '@bookmark/store/reducer';
 export { default as network } from '@network/store/reducer';
@@ -8,6 +8,6 @@ export { default as transactions } from '@transaction/store/reducer';
 export { default as appUpdates } from '@update/store/reducers/appUpdates';
 export { default as voting } from '@dpos/validator/store/reducers/voting';
 export { default as watchList } from '@dpos/validator/store/reducers/watchList';
-export { default as service } from './service';
+export { default as service } from '@wallet/store/reducers/service';
 export { default as loading } from 'src/modules/common/store/reducers/loading';
 export { account } from 'src/modules/account/store/reducer';
