+++ src/locales/en/common.json
@@ -103,6 +103,7 @@
   "Copied!": "Copied!",
   "Copy": "Copy",
   "Copy address": "Copy address",
+  "Copy link": "Copy link",
   "Copy public key": "Copy public key",
   "Copy to clipboard": "Copy to clipboard",
   "Count": "Count",
@@ -359,6 +360,9 @@
   "Report delegate misbehaviour": "Report delegate misbehaviour",
   "Report the error via email": "Report the error via email",
   "Reporter": "Reporter",
+  "Request LSK": "Request LSK",
+  "Request LSK from another Lisk user. Copy the link or scan the QR code to share your request.": "Request LSK from another Lisk user. Copy the link or scan the QR code to share your request.",
+  "Requested amount": "Requested amount",
   "Required": "Required",
   "Required signatures": "Required signatures",
   "Requirements": "Requirements",
@@ -412,6 +416,7 @@
   "Signature": "Signature",
   "Signed": "Signed",
   "Signed message": "Signed message",
+  "Simply scan the QR code using the Lisk Mobile app or any other QR code reader.": "Simply scan the QR code using the Lisk Mobile app or any other QR code reader.",
   "Since you want to reclaim your LSK on the new blockchain, you need to pay the fee from your new account.": "Since you want to reclaim your LSK on the new blockchain, you need to pay the fee from your new account.",
   "Slot status": "Slot status",
   "Something went wrong. Please try again": "Something went wrong. Please try again",
