+++ src/modules/wallet/components/request/request.js
@@ -4,9 +4,9 @@ import { maxMessageLength } from '@transaction/configuration/transactions';
 import { validateAmountFormat } from 'src/utils/validators';
 import { sizeOfString } from 'src/utils/helpers';
 import { Input, AutoResizeTextarea } from 'src/theme';
-import CircularProgress from 'src/theme/ProgressCircular/circularProgress';
+import CircularProgress from '@theme/ProgressCircular/circularProgress';
 import Converter from 'src/modules/common/components/converter';
-import Icon from 'src/theme/Icon';
+import Icon from '@theme/Icon';
 import i18n from 'src/utils/i18n/i18n';
 import RequestWrapper from './requestWrapper';
 import styles from './request.css';
