+++ src/modules/wallet/store/reducers/service.test.js
